requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
dataclasses-json==0.6.7
datasets==4.4.1
debugpy==1.8.18
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
executing==2.2.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.43.0
google-genai==1.55.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.7.0
jupyter_core==5.9.1
langchain==1.1.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.0
langchain-huggingface==1.2.0
langchain-text-splitters==1.0.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mmh3==4.1.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.3.5
orjson==3.11.5
ormsgpack==1.12.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pinecone==8.0.0
pinecone-client==6.0.0
pinecone-notebooks==0.1.1
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
pinecone-text==0.11.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uuid_utils==0.12.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
